requirements.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
Flask>=2.3.0
Flask-CORS>=4.0.0
Pillow>=11.0.0
pytesseract>=0.3.10
python-dateutil>=2.8.2
nltk>=3.8.1
requests>=2.31.0
Werkzeug>=2.3.0
bcrypt>=4.0.1
